+++ react/CurrentSellerContext.tsx
@@ -30,3 +30,8 @@ export const CurrentSellerProvider: FC<ProviderProps> = ({
     </CurrentSellerContext.Provider>
   )
 }
+
+export default {
+  CurrentSellerProvider,
+  useCurrentSeller,
+}