+++ react/SellerAddToCart.tsx
@@ -1,13 +1,14 @@
 import React from 'react'
 import { useCssHandles } from 'vtex.css-handles'
-import { useCurrentSeller } from './CurrentSellerContext'
 import { defineMessages } from 'react-intl'
 import { ExtensionPoint, useChildBlock } from 'vtex.render-runtime'
+
+import { useCurrentSeller } from './CurrentSellerContext'
 
 const SELLERS_CSS_HANDLES = ['sellerBuyContainer'] as const
 
 interface Props {
-  isOneClickBuy?: boolean,
+  isOneClickBuy?: boolean
 }
 
 const SellerAddToCart: StorefrontFunctionComponent<Props> = ({
@@ -16,13 +17,16 @@ const SellerAddToCart: StorefrontFunctionComponent<Props> = ({
   const { currentSeller } = useCurrentSeller()
   const handles = useCssHandles(SELLERS_CSS_HANDLES)
 
-  const useAddToCart = !!useChildBlock({ id: 'add-to-cart-button' });
+  const useAddToCart = !!useChildBlock({ id: 'add-to-cart-button' })
 
   return (
     <div
       className={`${handles.sellerBuyContainer} items-center tc br2 ph6 pv4 ma0 w-100-s w-20-m`}
     >
-      <ExtensionPoint id={useAddToCart ? 'add-to-cart-button' : 'buy-button'} {...{ isOneClickBuy, selectedSeller: currentSeller }} />
+      <ExtensionPoint
+        id={useAddToCart ? 'add-to-cart-button' : 'buy-button'}
+        {...{ isOneClickBuy, selectedSeller: currentSeller }}
+      />
     </div>
   )
 }
@@ -34,8 +38,8 @@ const messages = defineMessages({
   },
   isOneClickBuy: {
     defaultMessage: '',
-    id: 'admin/editor.seller-selector.oneClickBuy-title'
-  }
+    id: 'admin/editor.seller-selector.oneClickBuy-title',
+  },
 })
 
 SellerAddToCart.schema = {
@@ -46,9 +50,8 @@ SellerAddToCart.schema = {
       title: messages.isOneClickBuy.id,
       type: 'boolean',
       default: false,
-    }
-  }
+    },
+  },
 }
-
 
 export default SellerAddToCart