+++ react/SellerContext.tsx
@@ -39,8 +39,8 @@ interface SellerContextValue {
 const SellerContext = createContext<SellerContextValue>({
   sellerList: null,
   shippingQuotes: null,
-  setShippingQuotes: () => { },
-  limitShownShippingInformation: 3
+  setShippingQuotes: () => {},
+  limitShownShippingInformation: 3,
 })
 
 export const useSellerContext = (): SellerContextValue => {
@@ -55,4 +55,9 @@ export const SellerProvider: FC<ProviderProps> = ({ value, children }) => {
   return (
     <SellerContext.Provider value={value}>{children}</SellerContext.Provider>
   )
+}
+
+export default {
+  SellerProvider,
+  useSellerContext,
 }