+++ react/SellerShipping.tsx
@@ -19,23 +19,27 @@ const SellerShipping: StorefrontFunctionComponent = () => {
       className={`${handles.sellerShipping} items-center tc w-100-s w-20-m br2 ph6 pv4 ma0 w-100-s w-20-m`}
     >
       {shipping ? (
-        shipping.slas.slice(0, limitShownShippingInformation).map((sla, index: number) => (
-          <p key={index} className={`${handles.sellerShippingText} ma0`}>
-            <FormattedMessage
-              id="store/seller-list.shipping-estimate"
-              values={{
-                name: sla.name,
-                price: <FormattedCurrency value={sla.price / 100} />,
-                estimate: (
-                  <TranslateEstimate shippingEstimate={sla.shippingEstimate} />
-                ),
-              }}
-            />
-          </p>
-        ))
+        shipping.slas
+          .slice(0, limitShownShippingInformation)
+          .map((sla, index: number) => (
+            <p key={index} className={`${handles.sellerShippingText} ma0`}>
+              <FormattedMessage
+                id="store/seller-list.shipping-estimate"
+                values={{
+                  name: sla.name,
+                  price: <FormattedCurrency value={sla.price / 100} />,
+                  estimate: (
+                    <TranslateEstimate
+                      shippingEstimate={sla.shippingEstimate}
+                    />
+                  ),
+                }}
+              />
+            </p>
+          ))
       ) : (
-          <FormattedMessage id="store/seller-list.pending" />
-        )}
+        <FormattedMessage id="store/seller-list.pending" />
+      )}
     </div>
   )
 }