+++ react/typings/vtex.render-runtime.d.ts
@@ -2,6 +2,6 @@ import { ReactType } from 'react'
 
 declare module 'vtex.render-runtime' {
   export function useRuntime(): any
-  export function useChildBlock({ }): any
+  export function useChildBlock(param: object): any
   export const Link: ReactType
 }