+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Export `CurrentSellerProvider` and `useCurrentSeller` of `CurrentSellerContext`.
+- Export `SellerProvider` and `useSellerContext` of `SellerContext`.
+
 ## [0.1.1] - 2020-04-16
 
 ## [0.2.0] - 2020-04-15
@@ -16,22 +21,31 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.1.0] - 2020-03-20
 
 ## [0.1.4] - 2020-04-08
+
 ### Added
+
 - README adjustments
 
 ## [0.1.3] - 2020-04-08
+
 ### Added
+
 - README adjustments
 
 ## [0.0.2] - 2020-04-08
+
 ### Added
+
 - Full documentation and Css adjustments
 
 ## [0.0.2] - 2020-03-25
+
 ### Added
+
 - Added Simple documentation
 
 ## [0.0.2] - 2020-03-20
-### Added
-- CSS Handles to every component block
 
+### Added
+
+- CSS Handles to every component block